Issue 30: Clone does not support UNC path to repository
This is a bit of a hack and only fixes it for the repository, not the destination.
There are probably many other places where a UNC path will not work and this is
expected to be fixed in new version of msysgit.
It is worth doing this here however as the most likely place for a user to require a UNC path is when cloning.