Updated to the latest msysgit code base (1.6.3.2), which added something that might give some speed improvement. IIRC mmap support was added for 1.6.3.
The big change however was that I added the index caching that I mentioned a while back. It generates a cached index file for the last commit (and re-generates it if HEAD's sha1 does not match the cached file), which (I think) avoids acessing a bunch of files in the git db, and reduces a normal status query to reading two index files.

Myagi