Fixed Issue #226: tortoisegit is searching for .git share on network drives
Well, TortoiseGit searches for \\sambaserver\.git which is a *share* NOT a
directory. That's why logs on samba server a constantly filled with the errors like
in my first message. Also this causes a slite delay on Windows Explorer.

It would definetely work if the check is only done on directories, e.i.
\\server\sharename\.git or \\server\sharename\folder\.git but as I told
\\server\.git is a share and should not be checked everytime I access network server.

Signed-off-by: Frank Li <lznuaa@gmail.com>