In case the crash report dll can't be loaded, don't set a vectored exception handler
TortoiseGit(32).dll registers an exception handler when loaded, but does not remove the handler when the dll is unloaded. This means that if an application loads that dll (by using certain common shell APIs) and the dll gets unloaded soon after, then an exception happens in that application, that application crashes: the exception handler points to an address that doesn't exist anymore.

Based on TortoiseSVN revision #23222.

(fixes issue #1265)

Signed-off-by: Sven Strickroth <email@cs-ware.de>