Use libgit2 searchpath for finding config files
This also improves the case git.exe isn't on the path and system gitattributes are used (up to now libgit2 did not know where to find the attributes file).

This only works for use with git_repository.

This isn't used for TGitCache (GitIndex), because we need to know which files to use exactly (for caching purposes).

Signed-off-by: Sven Strickroth <email@cs-ware.de>