TortoiseGitMerge fails to restore Unicode characters
If the destination view is in ASCII and the source is not, we have to make sure that the text we copy from the other view can actually be saved in ASCII encoding.
If not, we have to change the destination view encoding to the same encoding as the source view encoding.

Based on TortoiseSVN revision 25161

Signed-off-by: Sup Yut Sum <ch3cooli@gmail.com>