Do not rely on unlocalized/hardcoded string for detecting detached HEAD on branch list
E.g. "git branch" using Git for Windows 2.4.5 on a German system shows:
  * (HEAD losgelöst bei 5a40e87)
    control-ref-prune-setting
    crlf
    development

Older git versions reported a name starting with "(no branch" or "(detached from ".

Also, branches starting with "(" are valid. Thus, we cannot use a starting "(" as a heuristic for detached HEAD and have to check more carefully.

Signed-off-by: Sven Strickroth <email@cs-ware.de>