Update TortoiseOverlays to 1.1.4-26626
The overlays on Win10 don't always show, because there are too many other overlay handlers installed. OneDrive installs 5 of them, and MS chose the name with a space in front of it to ensure those are loaded.
Unfortunately, the OneDrive overlays are not "good citizens" like Tortoise is so they always load *all* of those handlers even if there are many other handlers installed.
To ensure the Tortoise overlays are loaded, change the registration so that the Tortoise handler names start with two spaces, therefore placing them *before* the OneDrive handlers.

Based on TortoiseSVN rev. 26717, 26721, and 26722.

Signed-off-by: Sven Strickroth <email@cs-ware.de>