Allow to change file encoding
Use the encoding comboboxes at the bottom right to reload the file with the specified encoding if the ctrl key is pressed.
The behavior is left as is if the ctrl key is not pressed, i.e. the shown file is converted to the specified encoding.

Based on TortoiseSVN revision #26882 and #26883.

Signed-off-by: Sven Strickroth <email@cs-ware.de>