Fixed issue #2646: Library broken in Windows 10
The custom library doesn't work anymore on Win10. Use the default "Documents" library type instead for Win8 and later: In Win8 and later, the Explorer uses a ribbon UI and doesn't allow to add custom commands, which was the whole reason for the custom library type. And that means the "Documents" library type is sufficient for the Git library folder.

Based on TortoieSVN rev. 27029.

Signed-off-by: Sven Strickroth <email@cs-ware.de>