Watch registry key instead of re-reading it again and again
Use RegNotifyChangeKeyValue() to get a notification event whenever a registry value below HKCU\Software\TortoiseGit is changed. If a value has changed, re-read all the registry variables to ensure we use the latest ones. This is much better than re-reading the registry variables in specific time intervals because now we only re-read the values if really necessary, and do so immediately after a change and not have to wait until the timeout elapses.

Based on TortoiseSVN rev. 26729.

Signed-off-by: Sven Strickroth <email@cs-ware.de>