Optimize credential helper settings page
When not running with elevated permissions and a system helper was active, then just opening this settings page caused other settings to be ignored and a new elevated settings dialog to be started when clicking on OK or apply.

This commit fixes this issue and reduces the options when not running elevated.

Signed-off-by: Sven Strickroth <email@cs-ware.de>