Use "Application Registration" for optimizing protocol support
Protocols aren't <features> anymore because they are properly registered (even for Windows 7, which requires custom registry entries to work). The user is allowed to switch back and forth between programs that support the following protocols: git, github-windows, smartgit, x-github-client and git.

The website https://github.com does not render the proper links to Windows 7 clients.  Nothing to be done about it, they need to change that code and render proper links.  The link they send now to browsers running on windows 7 sends the users to a page to download their own "Github desktop" client (https://desktop.github.com/), but that client cannot be installed on Windows 7.

Signed-off-by: Guillaume Lachance <glachance@cmaisonneuve.qc.ca>
Signed-off-by: Sven Strickroth <email@cs-ware.de>