Improve performance
With a caveat: When fully staging/unstaging a file, I removed the trigger for a full refresh that I had added when clicking a checkbox. That means the group (Modified Files, Not Versioned Files...) and status (Added, Modified, Unknown...) will not get updated unless the user manually triggers a full refresh by pressing F5 (committing works without the full refresh, only the info displayed is affected). Automatically updating the group and status when clicking a checkbox without triggering a full refresh would require more rewriting. I added a note about that in the tip displayed when first enabling staging support.

As for supporting deleted files that had been staged, it's possible to do it in a half-baked way by simply removing the warning when staging support is enabled, but then the file groups, statuses and checkboxes would be incorrectly displayed and, worse, clicking the checkbox could do the wrong action. Doing it properly would also need more rewriting (in the "get working tree changes" code). So I left the warning as is.

Signed-off-by: Matheus Michels <63876508+tchesftw@users.noreply.github.com>
Signed-off-by: Sven Strickroth <email@cs-ware.de>