Implement expanding/collapsing nodes in the log tree on demand
This adds new commands in the context menu to collapse or expand nodes
of the log graph. If a node is collapsed, its parents are not shown up
to the next merge, fork, or labeled commit, similarly to a compressed
graph. Such nodes are displayed as hollow circles or rectangles.

The new commands support both normal and compressed graphs, allowing a
user to temporarily roll up branches that clutter the view or to expand
commit sets for inspecting them in detail.

(fixes issue #3787)

Signed-off-by: Igor Kudrin <ikudrin@accesssoftek.com>
Signed-off-by: Sven Strickroth <email@cs-ware.de>