Fixed issue #3940: restore scroll position of list
Change restoring the scroll position of the list of files in dialog "Commit" to be independent of GetTopIndex(), which only worked for lists with no selection. For lists with selected files a proper restoring was not possible due to lacking information about the first visible entry in the view. With GetOrigin() a solution is given, which works both for lists with non-selected or selected items.

Signed-off-by: Michael Bauer <meigelb_dev@outlook.com>
Signed-off-by: Sven Strickroth <email@cs-ware.de>