Fix possible crash when parent window is closed before push/pull
When a push, pull, or checkout dialog is launched from the context menu of an
explorer window, and the explorer window is subsequently closed before the
operation completes, TortoiseGitProc crashes. This can be annoying and
frustrating, as nothing happens afterward.

It appears that when trying to push/pull from the context menu, TortoiseGitProc
is called with the parameter /hwnd:0000000000XXXXXX. Then, if the window
associated with that hWnd has been closed, CWnd::FromHandle is called with an
invalid hWnd, which causes a crash later. So, it would be better to check if
the hWnd is still valid.

Signed-off-by: Ilya Lyakhovets <ilyalyakhovets@gmail.com>
Signed-off-by: Sven Strickroth <email@cs-ware.de>